add users page to cont3xt
put common user apis in common/user
