+++ cont3xt/vueapp/src/components/pages/Users.vue
@@ -1,14 +1,37 @@
 <template>
   <div class="container-fluid">
-    <UsersCommon />
+    <UsersCommon
+      v-if="getUser"
+      :roles="getRoles"
+      parent-app="Cont3xt"
+      @update-roles="updateRoles"
+      :current-user-id="getUser.userId"
+      @update-current-user="updateCurrentUser">
+    </UsersCommon>
   </div>
 </template>
 
 <script>
-import UsersCommon from '../../../../../common/Users';
+import { mapGetters } from 'vuex';
+
+import UsersCommon from '../../../../../common/vueapp/Users';
+import UserService from '@/components/services/UserService';
 
 export default {
   name: 'Users',
-  components: { UsersCommon }
+  components: { UsersCommon },
+  computed: {
+    ...mapGetters(['getUser', 'getRoles'])
+  },
+  methods: {
+    updateRoles () {
+      // NOTE: don't need to do anything with the data (the store does it)
+      UserService.getRoles();
+    },
+    updateCurrentUser () {
+      // NOTE: don't need to do anything with the data (the store does it)
+      UserService.getUser();
+    }
+  }
 };
 </script>
